+++ system/chat/opposed-chat.mjs
@@ -44,6 +44,8 @@ export class OPCard {
           targetScore: revisedtargetScore,
           rollResult: rollResult,
           cardType,
+          critChance: i.critChance,
+          fumbleChance: i.fumbleChance
         })
 
         i.rollResult = rollResult