+++ system/item/sheets/weapon-sheet.mjs
@@ -1,13 +1,16 @@
 import { AoVItemSheet } from "./base-item-sheet.mjs"
 import { AOVSelectLists } from "../../apps/select-lists.mjs"
+import { AOVActiveEffectSheet } from "../../sheets/aov-active-effect-sheet.mjs"
 
 export class AoVWeaponSheet extends AoVItemSheet {
   constructor(options = {}) {
-    super(options)
+    super(options);
+    this.#dragDrop = this.#createDragDropHandlers();
   }
 
   static DEFAULT_OPTIONS = {
     classes: ['weapon'],
+    dragDrop: [{ dragSelector: '[data-drag]', dropSelector: '.droppable' }],
     position: {
       width: 610,
       height: 670
@@ -19,6 +22,7 @@ export class AoVWeaponSheet extends AoVItemSheet {
     tabs: { template: 'systems/aov/templates/generic/tab-navigation.hbs' },
     details: { template: 'systems/aov/templates/item/weapon.detail.hbs' },
     description: { template: 'systems/aov/templates/item/item.description.hbs' },
+    effects: {template: 'systems/aov/templates/item/item.active-effect.hbs'},
     gmTab: { template: 'systems/aov/templates/item/item.gmtab.hbs' }
   }
 
@@ -50,7 +54,9 @@ export class AoVWeaponSheet extends AoVItemSheet {
     if (['missile','thrown'].includes(context.system.weaponType)) {
       context.isRange = true
     }
-
+    context.effects = AOVActiveEffectSheet.getItemEffectsFromSheet(this.document)
+    const changesActiveEffects = AOVActiveEffectSheet.getEffectChangesFromSheet(this.document)
+    context.effectChanges = changesActiveEffects.effectChanges
     return context
   }
 
@@ -58,6 +64,7 @@ export class AoVWeaponSheet extends AoVItemSheet {
   async _preparePartContext(partId, context) {
     switch (partId) {
       case 'details':
+      case 'effects':
         context.tab = context.tabs[partId];
         break;
       case 'description':
@@ -89,7 +96,7 @@ export class AoVWeaponSheet extends AoVItemSheet {
   _configureRenderOptions(options) {
     super._configureRenderOptions(options);
     //Only show GM tab if you are GM
-    options.parts = ['header', 'tabs', 'details','description'];
+    options.parts = ['header', 'tabs', 'details','effects','description'];
     if (game.user.isGM) {
         options.parts.push('gmTab');
     }
@@ -115,6 +122,10 @@ export class AoVWeaponSheet extends AoVItemSheet {
           tab.id = 'details';
           tab.label += 'details';
           break;
+        case 'effects':
+            tab.id = 'effects';
+            tab.label += 'effects';
+            break;
         case 'description':
           tab.id = 'description';
           tab.label += 'description';
@@ -133,6 +144,8 @@ export class AoVWeaponSheet extends AoVItemSheet {
 
   //Activate event listeners using the prepared sheet HTML
   _onRender(context, _options) {
+    this.#dragDrop.forEach((d) => d.bind(this.element))
+    AOVActiveEffectSheet.activateListeners(this)
     this.element.querySelectorAll('.weapon-type').forEach(n => n.addEventListener("change", this.#weaponType.bind(this)))
   }
 
@@ -158,6 +171,80 @@ export class AoVWeaponSheet extends AoVItemSheet {
     })
   }
 
+  // DragDrop
+  //
+  //
 
+  _canDragStart(selector) {
+    // game.user fetches the current user
+    return this.isEditable;
+  }
+
+  _canDragDrop(selector) {
+    // game.user fetches the current user
+    return this.isEditable;
+  }
+
+
+  _onDragStart(event) {
+    const li = event.currentTarget;
+    if ('link' in event.target.dataset) return;
+
+    let dragData = null;
+
+    // Active Effect
+    if (li.dataset.effectId) {
+      const effect = this.item.effects.get(li.dataset.effectId);
+      dragData = effect.toDragData();
+    }
+
+    if (!dragData) return;
+
+    // Set data transfer
+    event.dataTransfer.setData('text/plain', JSON.stringify(dragData));
+  }
+
+  _onDragOver(event) {}
+
+  //Handle the dropping of ActiveEffect data onto an Item Sheet
+  async _onDropActiveEffect(event, effect) {
+    let newEffect = effect.toObject();
+    newEffect.transfer = true
+    const item = this.document;
+    if ( !this.isEditable || !item.isOwner || (item === effect.parent) ) return null;
+    const result = await ActiveEffect.implementation.create(newEffect, {parent: item});
+    return result ?? null;
+  }
+
+   async _onDropItem(event, data) {
+    if (!this.item.isOwner) return false;
+  }
+
+   async _onDropFolder(event, data) {
+    if (!this.item.isOwner) return [];
+  }
+
+   get dragDrop() {
+    return this.#dragDrop;
+  }
+
+  // This is marked as private because there's no real need
+  // for subclasses or external hooks to mess with it directly
+  #dragDrop;
+
+   #createDragDropHandlers() {
+    return this.options.dragDrop.map((d) => {
+      d.permissions = {
+        dragstart: this._canDragStart.bind(this),
+        drop: this._canDragDrop.bind(this),
+      };
+      d.callbacks = {
+        dragstart: this._onDragStart.bind(this),
+        dragover: this._onDragOver.bind(this),
+        drop: this._onDrop.bind(this),
+      };
+      return new foundry.applications.ux.DragDrop.implementation(d);
+    });
+  }
 
 }